refactor: fecha y hora de dim_tiempo
1. **Migraciones (`dim_tiempo` y `fact_incidencias`)**: Se cambió la columna `tiempo_id` a tipo **`bigint`** (`unsignedBigInteger`) para soportar números de 14 dígitos sin causar desbordamiento de enteros (el límite máximo de un entero común de 32 bits es ~4.29 mil millones y `20260716025843` lo supera por mucho).
2. **Código ETL (`SyncFactIncidenciasJob.php`)**: Cambiamos el formato de `$tiempoId` de `YmdH` a **`YmdHis`** para que las marcas de tiempo tengan resolución de segundos completos.
3. **Base de Datos y Pruebas**:
   * Reconstruimos la base de datos local y sembramos los datos.
   * Agregamos 4 incidencias de prueba reales con minutos/segundos específicos en el historial.
   * Ejecutamos el ETL y verificamos que se registraran correctamente con su nuevo formato completo (ej. `20260714033015` en vez de `2026071403`).

+++ backend/api/database/migrations/olap/2026_06_23_000001_create_dim_tiempo_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
         DB::statement('CREATE SCHEMA IF NOT EXISTS metrics');
 
         Schema::create('metrics.dim_tiempo', function (Blueprint $table) {
-            $table->unsignedInteger('id')->primary(); // formato YYYYMMDDHH
+            $table->unsignedBigInteger('id')->primary(); // formato YYYYMMDDHHMMSS
             $table->timestamp('fecha');
             $table->unsignedSmallInteger('anio');
             $table->unsignedTinyInteger('mes');
